rename catalogs_safety_checks.py to catalogs.py to simplify file name and allow other methods in that file later
